Rename Sources, Sinks -> *Processors
Aligns with changes in the pipelines branch, to prevent merge conflicts.
