+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/connector/ReadFilesP.java
@@ -19,7 +19,7 @@ package com.hazelcast.jet.impl.connector;
 import com.hazelcast.jet.AbstractProcessor;
 import com.hazelcast.jet.Processor;
 import com.hazelcast.jet.ProcessorSupplier;
-import com.hazelcast.jet.processor.Sources;
+import com.hazelcast.jet.processor.SourceProcessors;
 
 import javax.annotation.Nonnull;
 import java.io.BufferedReader;
@@ -37,7 +37,7 @@ import java.util.stream.StreamSupport;
 import static com.hazelcast.jet.impl.util.ExceptionUtil.sneakyThrow;
 
 /**
- * @see Sources#readFiles(String, Charset, String)
+ * @see SourceProcessors#readFiles(String, Charset, String)
  */
 public class ReadFilesP extends AbstractProcessor {
 
@@ -96,7 +96,7 @@ public class ReadFilesP extends AbstractProcessor {
     }
 
     /**
-     * @see Sources#readFiles(String, Charset, String)
+     * @see SourceProcessors#readFiles(String, Charset, String)
      */
     public static ProcessorSupplier supplier(@Nonnull String directory, @Nonnull String charset, @Nonnull String glob) {
         return new ProcessorSupplier() {
