+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/stream/DistributedCollectors.java
@@ -63,8 +63,8 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collector;
 
-import static com.hazelcast.jet.processor.Sinks.writeCache;
-import static com.hazelcast.jet.processor.Sinks.writeMap;
+import static com.hazelcast.jet.processor.SinkProcessors.writeCache;
+import static com.hazelcast.jet.processor.SinkProcessors.writeMap;
 import static com.hazelcast.jet.Util.entry;
 
 /**
