+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/benchmark/WordCountTest.java
@@ -57,8 +57,8 @@ import static com.hazelcast.jet.Partitioner.HASH_CODE;
 import static com.hazelcast.jet.processor.Processors.flatMap;
 import static com.hazelcast.jet.processor.Processors.aggregateByKey;
 import static com.hazelcast.jet.processor.Processors.noop;
-import static com.hazelcast.jet.processor.Sources.readMap;
-import static com.hazelcast.jet.processor.Sinks.writeMap;
+import static com.hazelcast.jet.processor.SourceProcessors.readMap;
+import static com.hazelcast.jet.processor.SinkProcessors.writeMap;
 import static com.hazelcast.jet.Util.entry;
 import static com.hazelcast.jet.function.DistributedFunctions.entryKey;
 import static com.hazelcast.jet.function.DistributedFunctions.wholeItem;
