+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/connector/HazelcastConnectorIntegrationTest.java
@@ -42,14 +42,14 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.processor.Sinks.writeCache;
-import static com.hazelcast.jet.processor.Sinks.writeList;
-import static com.hazelcast.jet.processor.Sinks.writeMap;
-import static com.hazelcast.jet.processor.Sources.readCache;
-import static com.hazelcast.jet.processor.Sources.readList;
-import static com.hazelcast.jet.processor.Sources.readMap;
-import static com.hazelcast.jet.processor.Sources.streamCache;
-import static com.hazelcast.jet.processor.Sources.streamMap;
+import static com.hazelcast.jet.processor.SinkProcessors.writeCache;
+import static com.hazelcast.jet.processor.SinkProcessors.writeList;
+import static com.hazelcast.jet.processor.SinkProcessors.writeMap;
+import static com.hazelcast.jet.processor.SourceProcessors.readCache;
+import static com.hazelcast.jet.processor.SourceProcessors.readList;
+import static com.hazelcast.jet.processor.SourceProcessors.readMap;
+import static com.hazelcast.jet.processor.SourceProcessors.streamCache;
+import static com.hazelcast.jet.processor.SourceProcessors.streamMap;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.IntStream.range;
 import static org.junit.Assert.assertEquals;
