+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/connector/WriteBufferedPTest.java
@@ -29,7 +29,7 @@ import com.hazelcast.jet.ProcessorSupplier;
 import com.hazelcast.jet.Vertex;
 import com.hazelcast.jet.Watermark;
 import com.hazelcast.jet.impl.util.ArrayDequeInbox;
-import com.hazelcast.jet.processor.Sinks;
+import com.hazelcast.jet.processor.SinkProcessors;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.Before;
@@ -113,7 +113,7 @@ public class WriteBufferedPTest extends JetTestSupport {
 
     private ProcessorSupplier getLoggingBufferedWriter() {
         // returns a processor that will not write anywhere, just log the events instead
-        return Sinks.writeBuffered(
+        return SinkProcessors.writeBuffered(
                 idx -> {
                     events.add("new");
                     return null;
