+++ hazelcast-jet-hadoop/src/test/java/com/hazelcast/jet/impl/connector/hadoop/ReadHdfsPTest.java
@@ -60,7 +60,7 @@ import java.util.concurrent.Future;
 import java.util.stream.IntStream;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.processor.Sinks.writeList;
+import static com.hazelcast.jet.processor.SinkProcessors.writeList;
 import static com.hazelcast.jet.impl.util.Util.uncheckRun;
 import static com.hazelcast.jet.processor.HdfsProcessors.readHdfs;
 import static java.util.stream.IntStream.range;
