+++ hazelcast-jet-hadoop/src/test/java/com/hazelcast/jet/impl/connector/hadoop/WriteHdfsPTest.java
@@ -50,8 +50,8 @@ import java.util.concurrent.Future;
 import java.util.stream.IntStream;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.processor.Sources.readMap;
-import static com.hazelcast.jet.processor.Sinks.writeList;
+import static com.hazelcast.jet.processor.SourceProcessors.readMap;
+import static com.hazelcast.jet.processor.SinkProcessors.writeList;
 import static com.hazelcast.jet.processor.HdfsProcessors.readHdfs;
 import static com.hazelcast.jet.processor.HdfsProcessors.writeHdfs;
 import static java.util.stream.Collectors.toMap;
