+++ hazelcast-jet-kafka/src/test/java/com/hazelcast/jet/impl/connector/kafka/WriteKafkaPTest.java
@@ -36,7 +36,7 @@ import java.util.concurrent.Future;
 import java.util.stream.Collectors;
 
 import static com.hazelcast.jet.Edge.between;
-import static com.hazelcast.jet.processor.Sources.readMap;
+import static com.hazelcast.jet.processor.SourceProcessors.readMap;
 import static com.hazelcast.jet.processor.KafkaProcessors.writeKafka;
 import static java.util.stream.IntStream.range;
